27-06 01

+++ libft/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: llalba <llalba@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/20 11:43:47 by llalba            #+#    #+#             */
-/*   Updated: 2021/05/25 15:18:59 by llalba           ###   ########.fr       */
+/*   Updated: 2021/06/27 16:03:36 by llalba           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	size_t				i;
-	int					is_positive;
+	long long			is_positive;
 	unsigned long long	nbr;
 
 	i = 0;
@@ -35,5 +35,5 @@ int	ft_atoi(const char *str)
 		nbr = 10 * nbr + (str[i] - 48);
 		str++;
 	}
-	return (is_positive * nbr);
+	return (is_positive * (long long)nbr);
 }
